:recycle: getFormParts
